+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ## Changes since v7.9.0
 
+- [#3072](https://github.com/oauth2-proxy/oauth2-proxy/pull/3072) feat: support for multiple github orgs #3072 (@daniel-mersch)
+
 # V7.9.0
 
 ## Release Highlights